The good way to test the reactor shutdown procedures - got from /twisted/internet/test/test_time.py
